+++ includes/CodeImatic.pas
@@ -3,9 +3,10 @@ unit CodeImatic;
 interface
 
 Uses cmd;
+
   
-function codegen(aProject: string; aVariables: string; aWorkingdirectory: string;aOptions: string): Integer;
-function codegenex(aProject: string; aProjectconfig: string; aVariables: string; aWorkingdirectory: string;aOptions: string; aPreCompatible: boolean): Integer;
+function codegen(aProject: string;  aWorkingdirectory: string;aOptions: string): Integer;
+function codegenex(aProject: string;  aVariableCmdLineList: TVariableCmdLineList; aWorkingdirectory: string;aOptions: string; aPreCompatible: boolean): Integer;
 function GetcodegenPath: String;
 
 implementation
@@ -19,17 +20,22 @@ begin
     RaiseException(erCustomError, 'Environment variable "CODEIMATIC_CODEGEN" - codeimatic.codegen.exe path. cannot be found.')
 end;
 
-function codegen(aProject: string; aVariables: string; aWorkingdirectory: string;aOptions: string): Integer;
+function codegen(aProject: string; aWorkingdirectory: string;aOptions: string): Integer;
 begin
-  result := codegenex(aProject, '',aVariables,aWorkingdirectory,aOptions, false);
+  result := codegenex(aProject,  NIL, aWorkingdirectory,aOptions, false);
 end;
 
-function codegenex(aProject: string; aProjectconfig: string; aVariables: string; aWorkingdirectory: string;aOptions: string; aPreCompatible: boolean): Integer;
+function codegenex(aProject: string; aVariableCmdLineList: TVariableCmdLineList; aWorkingdirectory: string;aOptions: string; aPreCompatible: boolean): Integer;
 var
   SB: TStringBuilder;
+  lsVarCmdLines: string;
 begin
   Result := -1;
-  
+
+  lsVarCmdLines := '';
+ // If Assigned(aVariableCmdLineList) then
+  //  lsVarCmdLines := aVariableCmdLineList.ToVarCmdLines;
+
   if not Folder.Exists(GetcodegenPath) then 
     begin
       if not aPreCompatible then 
@@ -41,25 +47,24 @@ begin
   try
     SB:= TStringBuilder.Create;
 
+
     if not aPreCompatible then 
        SB.Append(GetcodegenPath + 'codeimatic.codegen.exe ')
     else
        SB.Append(GetcodegenPath + 'zcodegen.exe ');
     
-
+(*
     SB.Append('-project ' + aProject + ' ');
-
-    if Trim(aProjectConfig) <> '' then
-      SB.Append('-projectconfig' + aProjectConfig + ' ');
-   
 
     if not aPreCompatible then  SB.Append('-consoleoutputonly ');
 
     if aWorkingdirectory <> '' then 
       SB.Append('-workingdirectory '+  aWorkingdirectory+ ' ');   
 
-    if aVariables <> '' then 
-       SB.Append('-var ' + aVariables + ' ');
+
+    if lsVarCmdLines <> '' then
+       SB.Append(lsVarCmdLines + ' ');
+
 
     if aOptions <> '' then
       SB.Append(aOptions + ' ');   
@@ -67,9 +72,12 @@ begin
     Output.logformat('Running: %s', [SB.ToString]);
 
     result := Exec(sb.ToString);
+    *)
   finally
     SB.Free;
   end;
+ 
+   Result := 0;
 end;