+++ samples/Simple/test.pas
@@ -23,6 +23,7 @@ begin
   End;
 end;
 
+
 procedure RaiseExceptionTest;
 begin
   RaiseException(erCustomError, 'TEST');